requirements.txt
torch 
torchvision
numpy 
pathlib
argparse
scikit-learn
Pillow